A machine for parsing predicates.
It's not useful (or easy) to work with Surveyor's infix expressions directly.
Instead, parsing them into ASTs and letting the compiler deal with it is
a much easier way to do things.  The Ragel state machine in this commit
is a first attempt at dealing with that task.  (Because I don't want to
hand-code it.)

Right now, the machine only prints out status lines.  Future commits
will add AST construction and integration with Imperator::Parser, which
in turn will lead to construction of dependency-evaluation code in the
backends.